fixed ambiguous call to pow

+++ libmini/mini/minibase.h
@@ -105,8 +105,8 @@ inline double FSQR(const double x) {return(x*x);}
 #undef fpow
 #define fpow(x,y) pow((double)(x),(double)(y))
 
-inline float fround2(const float v,const unsigned int n=6) {return(ffloor(v*fpow(10.0f,n)+0.5f)/fpow(10.0f,n));}
-inline double round2(const double v,const unsigned int n=6) {return(floor(v*pow(10.0,n)+0.5)/pow(10.0,n));}
+inline float fround2(const float v,const unsigned int n=6) {return(ffloor(v*fpow(10.0f,(int)n)+0.5f)/fpow(10.0f,(int)n));}
+inline double round2(const double v,const unsigned int n=6) {return(floor(v*pow(10.0,(int)n)+0.5)/pow(10.0,(int)n));}
 
 #ifdef _MSC_VER
 #define strdup _strdup
